Refactored code to match latest create react app version
