+++ starter-code/.gitignore
@@ -1,7 +1,7 @@
 # See https://help.github.com/ignore-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+node_modules
 
 # testing
 /coverage
@@ -21,3 +21,5 @@ yarn-debug.log*
 yarn-error.log*
 
 .DS_Store
+package-lock.json
+yarn.lock
